fix(tests): finish moving code out of index, test main, fix resulting issues
