+++ run.mk
@@ -1,6 +1,6 @@
 export NODE_OPTIONS ?= --max-old-space-size=5500
 
-MOCHA_LONG_OPTS := $(MOCHA_OPTS) --timeout 5000
+MOCHA_LONG_OPTS := $(MOCHA_OPTS) --timeout 5000 --globals navigator
 
 SHELL := bash
 
@@ -11,6 +11,7 @@ local: build test-env
 test-env:
 	( export ISOLEX_HOME=$(ROOT_PATH)/docs; \
 	source $${ISOLEX_HOME}/isolex.env; \
+	rm $(TARGET_PATH)/isolex.pid; \
 	$(NODE_BIN)/nyc $(COVER_OPTS) \
 	  $(NODE_BIN)/mocha $(MOCHA_LONG_OPTS) $(TARGET_PATH)/test.js)
